WIP splitting things up in folders with own README’s because why not
