+++ tests/README.md
@@ -1,0 +1,23 @@
+[back to hapi-couchdb-account](..)
+
+# Testing
+
+Local setup
+
+```
+git clone git@github.com:hoodiehq/hapi-couchdb-account-api.git
+cd hapi-couchdb-account-api
+npm install
+```
+
+Run all tests and code style checks
+
+```
+npm test
+```
+
+Run all tests on file change
+
+```
+npm run test:watch
+```
