+++ tests/integration/fixtures/session-with-profile-response.json
