+++ tests/integration/utils/authorization-header-not-allowed-error.js
