+++ utils/session/delete.js
